requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
astor==0.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.61
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///D:/bld/cffi_1625835485017/work
cfgv==3.3.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
distlib==0.3.2
fastapi==0.66.0
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
future==0.18.2
gast==0.3.3
h11==0.12.0
identify==2.2.11
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
kiwisolver==1.3.1
lmdb==1.2.1
MarkupSafe==2.0.1
matplotlib==3.4.2
mccabe==0.6.1
networkx==2.5.1
nodeenv==1.6.0
numpy @ file:///D:/bld/numpy_1626648573548/work
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
paddleocr==2.0.6
paddlepaddle==2.1.1
pandas==1.3.0
Pillow==8.3.1
platformdirs==2.0.2
pre-commit==2.13.0
protobuf==3.17.3
pyclipper==1.3.0
pycodestyle==2.7.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pycryptodome==3.10.1
pydantic==1.8.2
pyflakes==2.3.1
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-Levenshtein==0.12.2
python-multipart==0.0.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.17.2
scipy==1.7.0
Shapely==1.7.1
shellcheck-py==0.7.2.1
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
starlette==0.14.2
tifffile==2021.7.2
toml==0.10.2
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
virtualenv==20.6.0
visualdl==2.2.0
Werkzeug==2.0.1
wincertstore==0.2
zipp==3.5.0
